% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggokabeito-package.R
\docType{package}
\name{ggokabeito-package}
\alias{ggokabeito}
\alias{ggokabeito-package}
\title{ggokabeito: 'Okabe-Ito' Scales for 'ggplot2' and 'ggraph'}
\description{
Discrete scales for the colorblind-friendly 'Okabe-Ito' palette, including 'color', 'fill', and 'edge_colour'. 'ggokabeito' provides 'ggplot2' and 'ggraph' scales to easily use the 'Okabe-Ito' palette in your data visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/malcolmbarrett/ggokabeito}
  \item \url{https://malcolmbarrett.github.io/ggokabeito/}
  \item Report bugs at \url{https://github.com/malcolmbarrett/ggokabeito/issues}
}

}
\author{
\strong{Maintainer}: Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})

}
\keyword{internal}
